refactor(client): colocate codex handler engines
